Fixed links to Arm XAR instruction documentation
The existing links to Arm's XAR page are broken, and Arm doesn't redirect to the correct ones. Their URL structure has changed (from /docs to now /documentation, and now including a date in the path, etc.)

Note that other links in the Readme to Arm's docs are likely broken for the same reason. I haven't checked the others.